# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Healthcare Recruiting ATS/CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates with filtering and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
        - name: search
          in: query
          schema:
            type: string
        - name: specialty
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: licenseType
          in: query
          schema:
            type: string
        - name: pipelineStage
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated candidate list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateListResponse"

    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a new candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidates/bulk-delete:
    post:
      operationId: bulkDeleteCandidates
      tags: [candidates]
      summary: Delete multiple candidates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkIdsInput"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /candidates/bulk-update:
    patch:
      operationId: bulkUpdateCandidates
      tags: [candidates]
      summary: Update multiple candidates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a single candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /candidates/{id}/stage:
    patch:
      operationId: updateCandidateStage
      tags: [candidates]
      summary: Update pipeline stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}/duplicate-action:
    post:
      operationId: handleDuplicateAction
      tags: [candidates]
      summary: Handle duplicate action (ignore, update, merge)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuplicateActionInput"
      responses:
        "200":
          description: Action applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # NOTES
  /candidates/{id}/notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"

    post:
      operationId: createNote
      tags: [notes]
      summary: Add a note to a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /candidates/{id}/notes/{noteId}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ACTIVITIES
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # AUTH
...
    Candidate:
      type: object
      required: [id, fullName, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        licenseType:
          type: ["string", "null"]
        licenseNumber:
          type: ["string", "null"]
        preferredLocation:
          type: ["string", "null"]
        availability:
          type: ["string", "null"]
        currentStatus:
          type: ["string", "null"]
        pipelineStage:
          type: ["string", "null"]
        recruiterNotes:
          type: ["string", "null"]
        resumeFile:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        isDuplicate:
          type: boolean
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        createdAt:
          type: string
        updatedAt:
          type: string
    CandidateInput:
      type: object
      required: [fullName]
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        position:
          type: string
        specialty:
          type: string
        experience:
          type: string
        licenseType:
          type: string
        licenseNumber:
          type: string
        preferredLocation:
          type: string
        availability:
          type: string
        currentStatus:
          type: string
        pipelineStage:
          type: string
        recruiterNotes:
          type: string
        source:
          type: string
    CandidateUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        position:
          type: string
        specialty:
          type: string
        experience:
          type: string
        licenseType:
          type: string
        licenseNumber:
          type: string
        preferredLocation:
          type: string
        availability:
          type: string
        currentStatus:
          type: string
        pipelineStage:
          type: string
        recruiterNotes:
          type: string
        source:
          type: string
    CandidateListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
    StageInput:
      type: object
      required: [stage]
      properties:
        stage:
          type: string
    BulkIdsInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
    BulkUpdateInput:
      type: object
      required: [ids, updates]
      properties:
        ids:
          type: array
          items:
            type: integer
        updates:
          type: object
          properties:
            pipelineStage:
              type: string
            currentStatus:
              type: string
            specialty:
              type: string
...
    DuplicateActionInput:
      type: object
      required: [action, targetId]
      properties:
        action:
          type: string
          enum: [ignore, update, merge]
        targetId:
          type: integer

    # NOTES
    Note:
      type: object
      required: [id, candidateId, content, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        content:
          type: string
        noteType:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    NoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        noteType:
          type: string
    NoteUpdate:
      type: object
      properties:
        content:
          type: string
        noteType:
          type: string

    # ACTIVITIES
...
    Tag:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: ["string", "null"]
...