# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Healthcare Recruiting ATS/CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { Tag } from './tag';

export interface Candidate {
  id: number;
  fullName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zipCode?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  experience?: string | null;
  /** @nullable */
  licenseType?: string | null;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  preferredLocation?: string | null;
  /** @nullable */
  availability?: string | null;
  /** @nullable */
  currentStatus?: string | null;
  /** @nullable */
  pipelineStage?: string | null;
  /** @nullable */
  recruiterNotes?: string | null;
  /** @nullable */
  resumeFile?: string | null;
  /** @nullable */
  source?: string | null;
  isDuplicate?: boolean;
  tags?: Tag[];
  createdAt: string;
  updatedAt: string;
}