# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Healthcare Recruiting ATS/CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface NoteUpdate {
  content?: string;
  noteType?: string;
}